fix: run newest debian
Ran into an issue with Debian-10 on macOS M1 (although not clear if the
os/arch has anything to do with this), where the latest version of
Parcel tried to pull in Rust-based components that relied on libc 2.29
or higher. Unfortunately Debian 10 ships with libc 2.28, so this crashed
at runtime.

This should be enough to fix it.